update `setNull` method NOT to pass specified javax.sql.Types as value
When setNull method was invoked, it was passing the second argument - int(javax.sql.Types) - to QueryInfo.
Currently, there is no way in QueryInfo whether the passed argument was from "setNull" or other method.
Also, in logging, it logs parameter value as "1=12"(for position 1 with VARCHAR) instead of indicating null.

In this commit, as a temporary workaround, ParameterSetOperation class will have two new methods: getParameterNameOrIndexAsString, getParameterValue.
When "setNull" is called, instead of passing int, it will return string "NULL" with corresponding Types name such as "VARCHAR". So that, in logging, it will simply write passed string rather than javax.sql.Types.

This logic should be revisited in next major version.

close #17