Proper before/after method callback for `ProxyDataSource#getConnection`
Update `ProxyDataSource` to use `ProxyDataSourceSupport`.
Prior to the change, it had a special treatment for `getConnection`
which does not provide a correct duration of the method.
This was done to always provide `ConnectionInfo` for both before/after
method callbacks.
However, since it is not correct, this commit fixes the order of method
and callback invocations.
For the before method callback, `ConnectionInfo` becomes `null` since
the target connection has not yet been acquired.

Resolve #100